insurance_recommendation_system/requirements.txt
gevent==22.10.2
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.10
pydantic==2.11.4
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
email-validator==2.2.0
python-dotenv==1.1.0
alembic==1.15.2
jinja2==3.1.2
aiosmtplib==4.0.1
httpx==0.27.0
pytest==7.4.0
locust==2.10.0
pytest-asyncio==0.19.0
pytest-cov==4.1.0